It's better to upload rpms from specific directory rather from /root
The main reason why rpms were moved from /root to a specific directory
is that /root is a generic directory and there could be stored any rpm.
So uploading rpms from /root will upload all rpms that are situated there,
even those that we would not like to or not aware of.